feat: 컴포넌트 분리 및 직접 영상 업로드 페이지
