+++ styles/upload/direct.module.css
@@ -1,0 +1,4 @@
+.wrap {
+  padding: 24px 16px;
+  margin-bottom: 10px;
+}
