+++ styles/upload/videoTitle.module.css
@@ -1,8 +1,4 @@
-.wrap {
-  padding: 24px 16px;
-  margin-bottom: 10px;
-}
-.link {
+.video_name {
   display: flex;
   flex-direction: column;
 }
@@ -22,20 +18,5 @@
   border-radius: 8px;
 }
 .text_input::placeholder {
-  color: #dadada;
-}
-
-/*description*/
-.description > textarea {
-  width: 100%;
-  border: 0.5px solid #c1c1c3;
-  border-radius: 8px;
-  font-family: pretendard;
-  padding: 15px 16px;
-  resize: none;
-  height: 120px;
-  font-size: 12px;
-}
-.description > textarea::placeholder {
   color: #dadada;
 }
